+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.5.1 - 2024-06-27
 
 * text-to-speach
+* focus on message input
+* resizing message input
 
 ## Version 0.5.0 - 2024-06-25